Winner Code! - Avoid Bullet Completed. [Todo #15]
Fix Spy Hidding  & Implement Edge Cells BFS.

Changes to be committed:
	modified:   Client/.idea/workspace.xml
	modified:   Client/Java/src/client/java/teamclient/AlliesInfo.java
	modified:   Client/Java/src/client/java/teamclient/EnemiesInfo.java
	modified:   Client/Java/src/client/java/teamclient/StaticsInfo.java
	modified:   Client/Java/src/client/java/teamclient/TeamClientAi.java
	new file:   Client/Java/src/client/java/teamclient/TiZiiClasses/TiZiiBullet.java
	renamed:    Client/Java/src/client/java/teamclient/TiZiiClasses/TiZiiCoord.java -> Client/Java/src/client/java/teamclient/TiZiiClasses/TiZiiCoords.java
	new file:   Client/Java/src/client/java/teamclient/TiZiiTest.java
	modified:   Client/Java/src/client/java/teamclient/TiZiiUtils.java
	modified:   README.md

+++ Client/Java/src/client/java/teamclient/TeamClientAi.java
@@ -1,10 +1,7 @@
 package client.java.teamclient;
 
 import client.java.communication.ClientGame;
-import client.java.teamclient.TiZiiClasses.DirectionScorePair;
-import client.java.teamclient.TiZiiClasses.DistanceDirectionPair;
-import client.java.teamclient.TiZiiClasses.StepsInDirection;
-import client.java.teamclient.TiZiiClasses.TiZiiCoord;
+import client.java.teamclient.TiZiiClasses.*;
 import common.board.Board;
 import common.board.Cell;
 import common.board.Direction;
@@ -16,6 +13,7 @@ import common.player.Spy;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.TreeMap;
 
 /**
  *  Class name:     TeamClientAi
@@ -43,17 +41,19 @@ public class TeamClientAi extends ClientGame {
         // updating.
         alliesInfo.updateAlliesInfo(getMyPlayers(), getOpponentPlayers(), getBullets()); // must use this before updating staticsBoard
         staticsInfo.updateStaticBoard(getGolds(), getMyPlayers());
-        enemiesInfo.updateEnemyBoard(getOpponentHunters(), getOpponentGoldMiners(), getOpponentSpies());
+        enemiesInfo.updateEnemyBoard(getOpponentHunters(), getOpponentGoldMiners(), getOpponentSpies(), getBullets());
         TiZiiUtils.update(alliesInfo);
 
         // logging
-        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.needed)       System.out.println(staticsInfo);
-        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.notNeeded)    System.out.println(enemiesInfo);
-        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.notNeeded)    TiZiiUtils.printBoard(gameBoard.getCells(), "Not Null Cells");
-        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.notNeeded)    TiZiiUtils.printBoard(staticsInfo.goldBFSTable, "Gold BFS TABLE");
-        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.needed)       TiZiiUtils.printGoldBfsDirections(staticsInfo.goldBFSTable, "Gold BFS Directions");
-        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.needed)       TiZiiUtils.printBoard(staticsInfo.discoveredAreas, "Discovered Areas.");
+        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.needed)          System.out.println(staticsInfo);
+        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.needed)          System.out.println(enemiesInfo);
+        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.notNeeded)       TiZiiUtils.printBoard(gameBoard.getCells(), "Not Null Cells");
+        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.notNeeded)       TiZiiUtils.printBoard(staticsInfo.goldBFSTable, "Gold BFS TABLE");
+        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.notNeeded)       TiZiiUtils.printGoldBfsDirections(staticsInfo.goldBFSTable, "Gold BFS Directions");
+        if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.notNeeded)       TiZiiUtils.printBoard(staticsInfo.discoveredAreas, "Discovered Areas.");
+	    if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.notNeeded)       TiZiiUtils.printBulletsHitTime(getBullets(), "Bullets Hit Time");
 
+        // Game Logic.
         ArrayList<GoldMiner> goldMiners = getMyGoldMiners();
         ArrayList<Hunter> hunters = getMyHunters();
         ArrayList<Spy> spies = getMySpies();
@@ -90,9 +90,12 @@ public class TeamClientAi extends ClientGame {
      * Runs per each miner.
      */
     private void minerLogic(GoldMiner miner) {
+	    if (enemiesInfo.isInRangeOfBullets(miner)){
+		    avoidBullets(miner); return;
+	    }
         Gold gold = gameBoard.getGold(miner.getCell());
         if (gold == null) {
-            TiZiiCoord u = new TiZiiCoord(miner.getCell());
+            TiZiiCoords u = new TiZiiCoords(miner.getCell());
 
             // if miner assigned to a gold and gold exists go there.
             Integer assignedGold = alliesInfo.assignedPlayerToGold.get(miner.getId());
@@ -128,11 +131,14 @@ public class TeamClientAi extends ClientGame {
      * Runs per each hunter.
      */
     private void hunterLogic(Hunter hunter) {
+	    if (enemiesInfo.isInRangeOfBullets(hunter)){
+		    avoidBullets(hunter); return;
+	    }
         if (hunter.getVisibleEnemy().size() > 0 && hunter.canAttack()
 		        && alliesInfo.noAlliesInsight(hunter)) {// can kill someone
             fire(hunter);
 	        for (Cell cell : hunter.getAheadCells()){
-		        alliesInfo.blockedCoords.add(new TiZiiCoord(cell));
+		        alliesInfo.blockedCoords.add(new TiZiiCoords(cell));
 	        }
         } else { // can not kill anyone
 	        // discovery move.
@@ -147,19 +153,68 @@ public class TeamClientAi extends ClientGame {
      * Runs per each spy.
      */
     private void spyLogic(Spy spy) {
-        if (enemiesInfo.isEnemyHunterNearby(spy.getCell())) spy.setHidden(true);
-        else spy.setHidden(false);
+	    System.out.println("Spy : " + spy.isHidden());
+	    if (enemiesInfo.isInRangeOfBullets(spy)){
+		    avoidBullets(spy); return;
+	    }
+        if (enemiesInfo.isEnemyHunterNearby(spy.getCell()) && !spy.isHidden()) {
+	        hide(spy); return;
+        }
+        else if (spy.isHidden()){
+	        show(spy); return;
+        }
+
         DirectionScorePair[] scorePair = alliesInfo.getDiscoveryMovementScores(spy);
         tiziiMove(spy, scorePair);
     }
 
-    private void tiziiMove(Player player, DirectionScorePair[] scorePair) {
-        Arrays.sort(scorePair); // TODO: Add a random value to not to take best direction
+	/**
+	 * @param player player that must avoid bullets
+	 */
+    private void avoidBullets(Player player){
+	    System.out.println("avoid Bullets " + player.getCell());
+	    TiZiiCoords tCoords = new TiZiiCoords(player.getCell());
+	    TreeMap<TiZiiBullet, Integer> bulletHitTimeIJ = enemiesInfo.bulletHitTime[tCoords.i][tCoords.j];
+
+	    for (TiZiiBullet tBullet :bulletHitTimeIJ.keySet()){
+			Integer time = bulletHitTimeIJ.get(tBullet);
+		    if (time == 1 && tBullet.direction != player.getMovementDirection()
+				    && tBullet.direction == TiZiiUtils.getReverseDirection(player.getMovementDirection())){
+
+			    if (canGo(player, player.getMovementDirection())) move(player);
+			    else                                              randomMove(player); // player is dead. :(
+		    }
+		    if (time == 2){
+			    // if Bullet Direction is Normal to Player Movement Direction.
+			    if (canGo(player, player.getMovementDirection()) && tBullet.direction != player.getMovementDirection()
+					    && tBullet.direction == TiZiiUtils.getReverseDirection(player.getMovementDirection())) {
+					move(player); return;
+			    }
+				// else turn player to a direction that he can go and its Normal to Bullet Direction.
+			    for (Direction direction : Direction.values()){
+				    if (canGo(player, direction) && tBullet.direction != direction
+						    && tBullet.direction == TiZiiUtils.getReverseDirection(direction)) {
+					    move(player); return;
+				    }
+			    }
+			    randomMove(player);
+			    // else player is dead :(
+		    }
+	    }
+    }
+
+	/**
+	 * Move Player To a Direction With Highest Score. (Includes Some Randomness).
+	 * @param player that needs to move.
+	 * @param scorePairs Array DirectionScorePair that Contain Score for Each Player.
+	 */
+    private void tiziiMove(Player player, DirectionScorePair[] scorePairs) {
+        Arrays.sort(scorePairs);
 
         boolean isFirst = true;
-        int rand = TiZiiUtils.getRandomNumber(5);
-        for (DirectionScorePair pair : scorePair) {
-            if (isFirst && rand < 1){ isFirst = false; continue; } isFirst = false; // Don't Use The Best answer sometimes.
+        int rand = TiZiiUtils.getRandomNumber(10);
+        for (DirectionScorePair pair : scorePairs) {
+            if (isFirst && rand < 1){ isFirst = false; continue; } isFirst = false; // Skip Best Score sometimes.
             if (canGo(player, pair.direction)) {
                 moveOrRotate(player, pair.direction);
                 StepsInDirection entry = alliesInfo.prevDirections.get(player.getId());
@@ -170,7 +225,7 @@ public class TeamClientAi extends ClientGame {
                 return;
             }
         }
-        for (DirectionScorePair pair : scorePair) {
+        for (DirectionScorePair pair : scorePairs) {
             if (canGo(player, pair.direction)) {
                 moveOrRotate(player, pair.direction);
                 StepsInDirection entry = alliesInfo.prevDirections.get(player.getId());
@@ -187,17 +242,17 @@ public class TeamClientAi extends ClientGame {
     private void moveOrRotate(Player player, Direction direction) {
         if (player.getMovementDirection().equals(direction) && canGo(player, direction)) {
             move(player);
-            alliesInfo.blockedCoords.add(new TiZiiCoord(player.getCell().getAdjacentCell(direction)));
+            alliesInfo.blockedCoords.add(new TiZiiCoords(player.getCell().getAdjacentCell(direction)));
         } else if (player.getMovementDirection().equals(direction) && !canGo(player, direction)){
 			randomMove(player);
         } else {
             rotate(player, direction);
-            alliesInfo.blockedCoords.add(new TiZiiCoord(player.getCell()));
+            alliesInfo.blockedCoords.add(new TiZiiCoords(player.getCell()));
         }
 
         // log
         if (TiZiiUtils.isLoggingEnabled && TiZiiUtils.notNeeded){
-            TiZiiCoord coord = new TiZiiCoord(player.getCell());
+            TiZiiCoords coord = new TiZiiCoords(player.getCell());
             if (player instanceof Spy)
                 System.out.println("Spy (" + coord.i + "," + coord.j + "): " + direction);
             if (player instanceof GoldMiner)
@@ -208,14 +263,14 @@ public class TeamClientAi extends ClientGame {
     private boolean canGo(Player player, Direction dir) {
         return player.getCell().getAdjacentCell(dir) != null
                 && player.getCell().getAdjacentCell(dir).isEmpty()
-                && !alliesInfo.blockedCoords.contains(new TiZiiCoord(player.getCell().getAdjacentCell(dir)));
+                && !alliesInfo.blockedCoords.contains(new TiZiiCoords(player.getCell().getAdjacentCell(dir)));
     }
 
     private void randomMove(Player player){
         int rand = TiZiiUtils.getRandomNumber(40);
         if (rand < 35 && canGo(player, player.getMovementDirection())) {
             move(player);
-	        TiZiiCoord nextCoords = new TiZiiCoord(player.getCell().getAdjacentCell(player.getMovementDirection()));
+	        TiZiiCoords nextCoords = new TiZiiCoords(player.getCell().getAdjacentCell(player.getMovementDirection()));
 	        alliesInfo.blockedCoords.add(nextCoords);
         } else {
 	        rand = TiZiiUtils.getRandomNumber(4);
